+++ _templates/project/app/routes.py
@@ -1,5 +1,5 @@
 from flask_unchained import (controller, resource, func, include, prefix,
-                             get, post, patch, put, rule)
+                             get, delete, post, patch, put, rule)
 
 from .views import SiteController